backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.9.1
psycopg[binary]==3.2.6
email-validator==2.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.0
apscheduler==3.10.4
stripe==12.0.1
jinja2==3.1.6
pytest==8.3.5
